// types/schema/contactSchema.ts
import { z } from "zod";

export const contactSchema = z.object({
  name: z.string().min(1, "Name is required").max(100),
  email: z.string().email("Invalid email"),
  subject: z.string().min(1, "Subject").max(1000),
  message: z
    .string()
    .min(10, "Message must be at least 10 characters")
    .max(2000),
});